requirements.txt
Flask
playwright
python-dotenv
gunicorn
supabase
cloudinary
requests
